change semantics of `Video::readFrame()`
Now it returns true if a new frame is available, false if not, and throws an error if something has gone wrong.

This is mostly so that we can poll `Video::Source` objects without having to wait for a new frame. It also  means you don't have to check for errors wherever `readframe()` is used.