rename classes and header file names for wheeled robots
`Motor` seemed a bit ambiguous to me, so I've renamed the base class to `Tank`. For the other classes I've dropped the Motor* prefix, because it's redundant now that we have a `GeNNRobotics::Robots` namespace.

I also renamed the `MotorI2C` class to `Norbot`. I thought I2C was slightly too vague a term and "Arduino" also seemed a bit ambiguous.

The aim is to ultimately stick all the Bebop header into to the robots folder once it's been tidied up.