Rename functions to do with loading snapshots
The terminology is a bit confusing as the images are not only being loaded, but also used to train the algorithm. Admittedly for the perfect memory case, these two are one and the same, but having "train" in the function name makes it explicit that this is what we're doing and is less confusing for e.g. infomax
Conflicts:
	navigation/visual_navigation_base.h