Workaround for bug in cv::DestroyWindow()
Apparently on Linux/Mac this function call often doesn't destroy the window, in which case we can get stuck in an infinite loop waiting for it to close. This hack makes sure that isOpen() returns false after close() is called so we can at least exit the program, even if the window may not actually close.