GPSReader: Use more sensible default device path on Linux
The problem with using /dev/ttyACM0 etc. as a name is that this can change depending on the order in which devices are plugged in. As we're also plugging in an Arduino this means that the device name for the GPS is not reliable. So let's use the device ID instead.

This does mean that for a different make/model of GPS we will need to update the path, but we can cross that bridge if/when we come to it.

Tag @tenxlenx