SFML::World: Change API to take event handler as argument
The current way of handling events is gross (and broken): mouse clicks are stored in the World object and can be retrieved later and if there's another event it's returned from the update() function. As events are handled in a loop (you have to poll through the event queue) it makes more sense to let the function take a callable thing to let the user do whatever they fancy with events.