Replace NMEA functions with an NMEAParser class
This also changes things so parsing errors are handled in GPSReader::read() (a warning is emitted) rather than passed up the call stack, so callers of this method don't need to check for errors anymore. Hopefully these errors should be rare (or non-existent) with the recent changes to the parsing and serial library code anyway.