Swap angles for navigation algos
This way round is more sensible: it means you're rotating anti-clockwise and so it's in the same direction as plots. It also aligns with what I've done in the python code.

I had to regenerate the navigation test data as the RIDFs will all now be reversed.

I only found one use of getHeading() that needed to be changed (in snapshot_bot).